==> 06.ipynb <==
1. データの読み込み

import math
import random
import time
import warnings
import numpy as np
import pandas as pd
import torch
import torch.nn as nn
import os
import transformers as T
from pathlib import Path
from sklearn.metrics import fbeta_score
from sklearn.model_selection import StratifiedKFold
from torch.utils.data import DataLoader, Dataset
import tqdm

warnings.filterwarnings("ignore")

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print(device)

def seed_torch(seed=42):
    random.seed(seed)
    os.environ["PYTHONHASHSEED"] = str(seed)
    np.random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    torch.cuda.manual_seed_all(seed)
    torch.backends.cudnn.deterministic = True

seed = 471
seed_torch(seed)

DATA_DIR = './dataset/data3'
OUTPUT_DIR = './result/result4'
if not os.path.isdir(OUTPUT_DIR):
    os.makedirs(OUTPUT_DIR)

train = pd.read_csv(DATA_DIR  +"/train.csv",  index_col=0)
test = pd.read_csv(DATA_DIR + "/test.csv", index_col=0)
sub = pd.read_csv(DATA_DIR + "/sample_submit.csv", header=None)
sub.columns = ["id", "judgement"]

train.head()

from imblearn.under_sampling import RandomUnderSampler
positive_count_train = train['judgement'].sum()
rus = RandomUnderSampler(random_state=71)
X, y = train.drop(columns=['title', 'abstract', 'judgement', 'title_token', 'abstract_token']), train['judgement']
X_resample, y_resample = rus.fit_resample(X,y)
train_resample = pd.concat([X_resample, y_resample], axis=1)
X_test = test.drop(columns=['title', 'abstract', 'title_token', 'abstract_token'])

from pycaret.classification import *
exp1 = setup(train_resample, target = 'judgement')

from sklearn.metrics import fbeta_score
def fbeta_score(x, y):
    score = fbeta_score(x, y, beta=7.0)
    return score
add_metric('fbeta', 'fbeta', fbeta_score, target = 'pred_proba', greater_is_better=True)
compare_models()

et = create_model('et')
tuned_et = tune_model(et)

et_predict = predict_model(tuned_et, X_test)

sub = pd.read_csv(DATA_DIR + "/sample_submit.csv", header = None)
sub.columns = ["id", "judgement"]
sub["judgement"] = et_predict['Label'].values

#保存先のディレクトリ作成
save_dir = "./result/result4"
if not os.path.isdir(save_dir):
    os.makedirs(save_dir)
sub.to_csv(save_dir + '/submission.csv', header=None, index=None)

evaluate_model(tuned_ridge)

sub = pd.read_csv(DATA_DIR + "/sample_submit.csv", header=None)
